token_dealer/requirements.txt
pyjwt
pyzmq